Fixed checkstyle errors.
