+++ src/test/java/dolla/command/modify/PartialModifyEntryCommandTest.java
@@ -1,6 +1,6 @@
 package dolla.command.modify;
 
-import dolla.DollaDataStubs.DollaDataEntryStub1;
+import dolla.dolladatastub.DollaDataEntryStub1;
 import dolla.ModeStringList;
 import dolla.model.DollaData;
 import dolla.model.Entry;
