+++ src/test/java/dolla/command/view/ViewDateCommandTest.java
@@ -10,7 +10,7 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 public class ViewDateCommandTest {
 
     @Test
-    public void ViewDateCommandInfo() {
+    public void viewDateCommandInfo() {
         LocalDate inputDate = LocalDate.parse("2001-12-03");
         Command newViewTodayCommand = new ViewDateCommand(inputDate);
         String expected = Time.dateToString(inputDate) + " " + Time.dateToString(inputDate);
