+++ src/test/java/dolla/dolladatastub/DollaDataLimitStub1.java
@@ -1,4 +1,4 @@
-package dolla.DollaDataStubs;
+package dolla.dolladatastub;
 
 import dolla.model.BillList;
 import dolla.model.DebtList;
@@ -10,15 +10,14 @@ import dolla.model.ShortcutList;
 
 import java.util.ArrayList;
 
-public class DollaDataEntryStub1 extends DollaData implements EntryArrayListInterface {
+public class DollaDataLimitStub1 extends DollaData implements LimitArrayListInterface {
 
     /**
      * Creates an instance of DollaData to store and manipulate data.
      */
-    public DollaDataEntryStub1() {
-        this.mode = MODE_ENTRY;
-        this.entryList = new EntryList(createNewEntryArrayList());
-        this.limitList = new LimitList(new ArrayList<Record>());
+    public DollaDataLimitStub1() {
+        this.entryList = new EntryList(new ArrayList<Record>());
+        this.limitList = new LimitList(createNewLimitArrayList());
         this.debtList = new DebtList(new ArrayList<Record>());
         this.shortcutList = new ShortcutList(new ArrayList<Record>());
         this.billList = new BillList(new ArrayList<Record>());
